test(api): consolidate ResponseEntity assertions and merge JwtAuthenticationIT
Use single recursive comparison with ResponseEntity.ok() instead of
split status + body assertions across 6 controller tests. Move 4
JwtAuthenticationIT tests into JwtToAuthenticatedUserConverterTest
unit test and delete the misplaced integration test file.
