+++ apps/api/main/src/test/java/io/stablepay/api/application/web/controller/AdminDlqControllerTest.java
@@ -30,7 +30,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 
 @ExtendWith(MockitoExtension.class)
 class AdminDlqControllerTest {
@@ -108,9 +108,8 @@ class AdminDlqControllerTest {
       var actual = controller.findById(SOME_DLQ_ID.value().toString(), SOME_ADMIN_USER);
 
       // then
-      assertThat(actual.getStatusCode()).isEqualTo(HttpStatus.OK);
-      var expected = mapper.toDto(SOME_DLQ_EVENT);
-      assertThat(actual.getBody()).usingRecursiveComparison().isEqualTo(expected);
+      var expected = ResponseEntity.ok(mapper.toDto(SOME_DLQ_EVENT));
+      assertThat(actual).usingRecursiveComparison().isEqualTo(expected);
     }
 
     @Test
